fix(utils): fixed gen detection in detectGenFromFormat
there's a format called gen8350cup (thanks brokenmotor for the
Teambuilder dump!), which resulted in gen 8350 after running it through
the GenFormatRegex. obviously there aren't any entries in gen 8350
(don't even think Game Freak thought that far), which caused the Calcdex
to silently crash when calling getTeambuilderPresets. the fix entails
only parsing a single digit after the gen prefix now, which is fine,
until gen 10 comes. but we'll cross that bridge when we get there...
LOL.